refactor: 위치 타입 전역으로 관리 [FE-71]
