--- src/features/map/map.types.ts
@@ -1,6 +1,0 @@
-interface LatLng {
-  latitude: number | undefined;
-  longitude: number | undefined;
-}
-
-export type { LatLng };
